// _web_legacy/src/api/model/contentBlockResponse.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Prospero API
 * OpenAPI spec version: 0.1.0
 */
import type { SentenceResponse } from './sentenceResponse';

export interface ContentBlockResponse {
  block_type: string;
  html: string;
  id: string;
  /** @nullable */
  paragraph_number?: number | null;
  position: number;
  sentences: SentenceResponse[];
}