// _web_legacy/src/api/model/sentenceResponse.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Prospero API
 * OpenAPI spec version: 0.1.0
 */
import type { PageMarkerResponse } from './pageMarkerResponse';

export interface SentenceResponse {
  html: string;
  id: string;
  page_markers: PageMarkerResponse[];
  position: number;
  /** @nullable */
  sentence_number?: number | null;
  text: string;
}